requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.13.1
reportlab==4.0.8
Pillow==10.1.0
python-dotenv==1.0.0
qrcode[pil]==7.4.2
openpyxl==3.1.2
pandas==2.1.4
